+++ Sai Nyayapati/NewCode/UARTCode/UARTCode/uart.h
@@ -14,8 +14,8 @@
 void usart_init(uint32_t baudRate);
 void usart_transmit_byte(char byte);
 void usart_transmit_array(char*msg);
-void usart_voltage(int voltage);
-void usart_current(int current);
+void usart_voltage(uint16_t voltage);
+void usart_current(uint16_t current);
 void usart_power(uint16_t power);
